:truck: Move test projects into subfolder test/../golden
