+++ packages/riverpod_graph/test/integration/addition/golden/lib/base_providers.dart
