+++ packages/riverpod_graph/test/integration/addition/golden/lib/combined_provider.dart
