+++ packages/riverpod_graph/test/integration/addition/golden/pubspec.yaml
@@ -7,4 +7,4 @@ environment:
 
 dependencies:
   riverpod:
-    path: ../../../riverpod
+    path: ../../../../../riverpod
