+++ packages/riverpod_graph/test/integration/consumer_widget/golden/lib/main.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+
+/// Counter provider.
+final counterProvider = StateProvider((ref) => 0);
+
+/// Counter widget.
+class CounterWidget extends ConsumerWidget {
+  /// Counter widget.
+  const CounterWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return Scaffold(
+      body: Consumer(builder: (context, ref, _) {
+        final count = ref.watch(counterProvider.state).state;
+        return Text('$count');
+      }),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () => ref.read(counterProvider.state).state++,
+        child: const Icon(Icons.add),
+      ),
+    );
+  }
+}
