+++ packages/riverpod_graph/test/integration/consumer_widget/golden/pubspec.yaml
@@ -13,26 +13,26 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_riverpod:
-    path: ../../../../packages/flutter_riverpod
+    path: ../../../../../../packages/flutter_riverpod
   hooks_riverpod:
-    path: ../../../../packages/hooks_riverpod
+    path: ../../../../../../packages/hooks_riverpod
   riverpod:
-    path: ../../../../packages/riverpod
+    path: ../../../../../../packages/riverpod
 
 dev_dependencies:
   custom_lint: ^0.0.7
   riverpod_lint:
-    path: ../../../../packages/riverpod_lint
+    path: ../../../../../../packages/riverpod_lint
   flutter_test:
     sdk: flutter
 
 dependency_overrides:
   flutter_riverpod:
-    path: ../../../../packages/flutter_riverpod
+    path: ../../../../../../packages/flutter_riverpod
   hooks_riverpod:
-    path: ../../../../packages/hooks_riverpod
+    path: ../../../../../../packages/hooks_riverpod
   riverpod:
-    path: ../../../../packages/riverpod
+    path: ../../../../../../packages/riverpod
 
 flutter:
   uses-material-design: true
